change some parameters in gfs_pio_gets(), gfs_pio_readline() and gfs_pio_readdelim() from int to size_t, because size_t is already used in other place in <gfarm/gfs.h>, so no need to worry about include extra header like <stddef.h>. pointed out by ikegami-san.
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@1140 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ include/gfarm/gfs.h
@@ -93,11 +93,12 @@ int gfs_pio_getc(GFS_File);
 int gfs_pio_ungetc(GFS_File, int);
 char *gfs_pio_putc(GFS_File, int);
 char *gfs_pio_puts(GFS_File, const char *);
-char *gfs_pio_gets(GFS_File, char *, int);
+char *gfs_pio_gets(GFS_File, char *, size_t);
 char *gfs_pio_getline(GFS_File, char *, size_t, int *);
 char *gfs_pio_putline(GFS_File, const char *);
-char *gfs_pio_readline(GFS_File, char **, int *, int *);
-char *gfs_pio_readdelim(GFS_File, char **, int *, int *, const char *, int);
+char *gfs_pio_readline(GFS_File, char **, size_t *, size_t *);
+char *gfs_pio_readdelim(GFS_File, char **, size_t *, size_t *,
+	const char *, size_t);
 
 /*
  *  For legacy code
